Testing how to map fist to palm open of hand using distance
testing out how to map hand opening and closing in a more controlled way, eg fist to palm open, that allows a continuous cycle between the two states. I did this by using the distance between the wrist x and y points and the middle finger tip x and y points.

This is important for my idea as I don't want it to just jump between things and instead morph smoothly between states.

I tested it out using it to morph between blue and pink with a lerp Color function. Making it blue if palm open and pink if closed.