initial tests with mapping the hand opening to a flower opening
Swapping out the lerpColor and mapping the same wrist to middle finger tip distance to the frame number when loading images.

eg. when the fist is closed, distance is small and the frame =1. when the palm is open fully, frame = 10.

this creates an animation that follows the hand opening and closing. it works both ways which is cool, giving the effect of the hand controlling the growth and closure of the flower.

This is just a quick test I made from a flower animation I created in blender and the video will be adapted to a more complex flower scene.