Adding in static grass and animated clouds
Adding in static grass in the foreground to make the flowers look like they belong. This was hand drawn in krita.

Adding in rolling clouds animation in the background. I precomposed 3 clouds together as one image to save processing power and so I only needed to load one image. I then used the X position of the cloud image to animate it, mapping the frameRate to the X position to get it to move. I used two instances of the image, as one instance look too sparse, The second cloud is staggered so that it starts when the first lot reach halfway through the sketch.

I wanted to add some horizontal movement to the scene and give the scene some life to match the growth interaction of the flowers and hand.

The clouds were drawn in krita and then comped into one image in photoshop.