Mapping the left hand to make the scene day or night
Further exploring the idea of sped up time with the left hand controlling the colour of the background, to make it either day or night. You can either quickly change from day to night or slowly open the hand to make it change slower.

Using lerpColor to do this and the using the same code to map the distance between the wrist and middelfinger tip to control a variable. This time being the value that morphs between two colours with lerp Color. Left fist closed = day time with the variable set to 0, meaning blue. palm open= night, set to 1, navy.

Also making it so the flower growth is only mapped to the right hand, giving distinct controls between the two hands. Making it it feel more interactable and controlled.