Adding in texture
Adding in textured overlays to give the sketch more character. I felt it looked a little too clean and digital for my liking.

I added in some scans of risograph  printed textures I had and used the tint function to adjust their opacity.

This added in some subtle ink on paper textures that add to the illustrated style.