Adding in a butterfly that is mapped to face position
Mapping a butterfly animation to follow where the face is on the X axis by mapping the face to the X location of the image. I chose not to map the Y location of the butterfly to the face as it got too chaotic and distracted from the rest of the scene. Adding it to only the X axis makes it subtle enough to not disrupt the scene.

Using the same code I used for the smoothing the hand mapping to make the animation smoother when it follows the face. Using the lerp Function.

Also making it so that the butterfly faces the same way the face moves, eg if the face moves to the right, the butterfly faces right and moves to the right. This made it look more natural, and made it feel more interactive.